replaced npm with yarn
npm is hitting some sort of circular logic on install. yarn does not. so yarn it is

still not sure why images aren't being served.

persistent node_modules was solved as well.  you have to add an exception in the volume mount in the docker-compose.

I changed the directory, and I might do it again, but its fine for now